explicit cast

+++ lib/hcrypto/aes.c
@@ -90,7 +90,7 @@ AES_cbc_encrypt(const unsigned char *in, unsigned char *out,
 	    out += AES_BLOCK_SIZE;
 	}
 	if (size) {
-	    for (i = 0; i < size; i++)
+	    for (i = 0; (unsigned long)i < size; i++)
 		tmp[i] = in[i] ^ iv[i];
 	    for (i = size; i < AES_BLOCK_SIZE; i++)
 		tmp[i] = iv[i];
@@ -111,7 +111,7 @@ AES_cbc_encrypt(const unsigned char *in, unsigned char *out,
 	if (size) {
 	    memcpy(tmp, in, AES_BLOCK_SIZE);
 	    AES_decrypt(tmp, out, key);
-	    for (i = 0; i < size; i++)
+	    for (i = 0; (unsigned long)i < size; i++)
 		out[i] ^= iv[i];
 	    memcpy(iv, tmp, AES_BLOCK_SIZE);
 	}
