+++ lib/krb5/recvauth.c
@@ -137,7 +137,7 @@ krb5_recvauth_match_version(krb5_context context,
 	}
 	len = ntohl(len);
 	if (len != sizeof(her_version)
-	    || krb5_net_read (context, p_fd, her_version, len) != len
+	    || (size_t)krb5_net_read (context, p_fd, her_version, len) != len
 	    || strncmp (version, her_version, len)) {
 	    repl = 1;
 	    krb5_net_write (context, p_fd, &repl, 1);
@@ -166,7 +166,7 @@ krb5_recvauth_match_version(krb5_context context,
 	krb5_net_write (context, p_fd, &repl, 1);
 	return krb5_enomem(context);
     }
-    if (krb5_net_read (context, p_fd, her_appl_version, len) != len
+    if ((size_t)krb5_net_read (context, p_fd, her_appl_version, len) != len
 	|| !(*match_appl_version)(match_data, her_appl_version)) {
 	repl = 2;
 	krb5_net_write (context, p_fd, &repl, 1);
